Harden strunredis.sh by not using a local TCP port
Your script is very helpful. However, it is not secure in a multiuser host. The script opens up a TCP port bound to the loopback device, which IMHO means that (at least without some ip[6]tables hardening) anyone on localhost can use the TCP port without having the cert and/or private key (in case of two-way trust TLS, which we use. `cert` and `key` in `stunnel`.). This might not seem relevant to (for example) a vagrant dev VM, but it's easy to avoid: just use a unix domain socket. `stunnel` as well as `redis-cli` supports it. **As a bonus:** the unix domain socket performs better as well.
Sidenote: `stunnel` removes the socket file automatically, with a clean exit. Please judge yourself if a more resilient cleanup is needed (I don't think so, but could be wrong).